move activity components to components folder
