+++ src/components/Activity/ActivityAddress.tsx
@@ -1,5 +1,5 @@
 import { Text } from '@chakra-ui/react';
-import useDisplayName from '../../../hooks/utils/useDisplayName';
+import useDisplayName from '../../hooks/utils/useDisplayName';
 
 export function ActivityAddress({ address, addComma }: { address: string; addComma: boolean }) {
   const { displayName } = useDisplayName(address);
